+++ CHANGELOG.md
@@ -9,9 +9,11 @@ The changelog is available [on GitHub][2].
   Support GHC-9.0. Update older GHC's bounds.
 * [#30](https://github.com/kowainik/slist/issues/30):
   Add the `cons` function.
-* [#29](https://github.com/kowainik/slist/issues/45):
+* [#29](https://github.com/kowainik/slist/issues/29):
   Add `Slist.Maybe` module with `maybeToSlist`, `slistToMaybe`, `catMaybes`,
   `mapMaybe`, `slistWith` functions.
+* [#24](https://github.com/kowainik/slist/issues/24):
+  Add `chunksOf` for `Slist` and `listChunksOf` for ordinary lists.
 * Move the `Slist` data type into the separate `Slist.Type` module.
 
 ## 0.1.1.0 — Apr 18, 2020